LFS OK. LFS and Flash diver commented

+++ Core/Inc/OS/OS_FS/lfs_flash.h
@@ -15,9 +15,9 @@
  *********************************************/
 
 #define SECTOR_SIZE					(128*1024)
-#define LFS_BASE_ADDR				&__lfs_start
-#define LFS_END_ADDR				&__lfs_end
-#define LFS_BASE_SIZE				&_LFS_SIZE
+#define LFS_BASE_ADDR				((uint32_t)&__lfs_start)
+#define LFS_END_ADDR				((uint32_t)&__lfs_end)
+#define LFS_BASE_SIZE				((uint32_t)&_LFS_SIZE)
 
 /**********************************************
  * PUBLIC VARIABLES
